+++ R/MicrobiomeR-package.R
@@ -23,5 +23,5 @@
 #'
 #' @author Rob Gilmore and Shaurita Hutchins
 #' @name MicrobiomeR
-#' @docType package
-NULL
+#' @keywords internal
+"_PACKAGE"